feat: Switch from k8s-tpl to Helm
Since Helm is a more idiomatic tool, it makes sense to use it if
possible. This commit shows it is indeed possible, and not entirely
horrible. Having to define a Chart for any local deployments in order to
use the templating is a bit verbose, but otherwise it works quite well.