feat: Encode deployment in a Helm chart
This dramatically simplifies the deployment steps, and the README has
been restructured to account for the fact that configuration now
accounts for most of the interaction.

The main unpleasantness is the need to keep the cert-manager CRDs in the
chart. For whatever reason, the CRDs were not deployed from the sub-
chart (even with includeCRDs=true), and this was the easiest way to work
around it.